Fix implicit conversion to int
Bug: T351312
(cherry picked from commit 3fc8f0a129eb8336daa76d5bfb63f96d316f0e35)

+++ includes/ParserFunctions.php
@@ -788,7 +788,7 @@ class ParserFunctions {
 		// Precompute limit to avoid generating enormous string:
 		$diff = mb_strlen( $inReplaceTo ) - mb_strlen( $inReplaceFrom );
 		if ( $diff > 0 ) {
-			$limit = ( ( $wgPFStringLengthLimit - mb_strlen( $inStr ) ) / $diff ) + 1;
+			$limit = (int)( ( $wgPFStringLengthLimit - mb_strlen( $inStr ) ) / $diff ) + 1;
 		} else {
 			$limit = -1;
 		}
